back-end/pyproject.toml
[project]
name = "wolliz-backend"
version = "0.1.0"
description = "API do marketplace imobiliário Wolliz"
requires-python = ">=3.12"

dependencies = [
    # ─── Django ───────────────────────────────────────────────
    "django>=5.1,<5.2",
    "django-ninja>=1.3,<2.0",
    "django-ninja-jwt>=5.3",

    # ─── Database ─────────────────────────────────────────────
    "psycopg2-binary>=2.9",

    # ─── Cache ────────────────────────────────────────────────
    "django-redis>=5.4",
    "redis>=5.0",

    # ─── Storage (MinIO / S3) ─────────────────────────────────
    "django-storages[s3]>=1.14",
    "boto3>=1.35",

    # ─── Server ───────────────────────────────────────────────
    "uvicorn[standard]>=0.34",
    "gunicorn>=23.0",

    # ─── CORS ─────────────────────────────────────────────────
    "django-cors-headers>=4.6",

    # ─── Utilities ────────────────────────────────────────────
    "python-decouple>=3.8",
    "pillow>=11.0",
    "email-validator>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "django-debug-toolbar>=4.4",
]
